Clean up the code a bit; fix some bugs in the CPU
This commit moves a lot of helper functions like split/concat into another
file to make it more readable; same with the Register class.

Additionally, we use named constants instead of plain values for
certain things (e.g. exception vectors).

The PBR is now stored on native mode exceptions (as it should be).

Don't use pointers into memory; all memory access should go through the
bus, using the methods it provides.

The PC is now updated to the next instruction BEFORE executing the
current instruction; this matches the actual behavior of the 65C816.